# improve performance of `unstructured_3d_to_1d_curve` for linear `P4estMesh`es in 3D 