With this PR, I get (after compilation)

```julia
julia> using Revise; using Trixi, Plots

julia> begin
       equations = CompressibleEulerEquations3D(1.4)
       solver = DGSEM(polydeg = 3, surface_flux = flux_lax_friedrichs)
       coordinates_min = (-1.0, -1.0, -1.0)
       coordinates_max = (+1.0, +1.0, +1.0)
       initial_refinement_level = 4

       mesh_tree = TreeMesh(coordinates_min, coordinates_max;
                            n_cells_max = 10^6, initial_refinement_level)
       trees_per_dimension = (1, 1, 1)
       mesh_t8code = T8codeMesh(trees_per_dimension; polydeg = 1,
                                coordinates_min, coordinates_max, initial_refinement_level)
       mesh_p4est = P4estMesh(trees_per_dimension; polydeg = 1,
                              coordinates_min, coordinates_max, initial_refinement_level)
       cells_per_dimension = (2, 2, 2) .^ initial_refinement_level
       mesh_structured = StructuredMesh(cells_per_dimension, coordinates_min, coordinates_max)

       ic = initial_condition_constant
       ode_tree = let mesh = mesh_tree
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_t8code = let mesh = mesh_t8code
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_p4est = let mesh = mesh_p4est
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_structured = let mesh = mesh_structured
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       x = range(0, 1, length = 1000)
       curve = vcat(x', x', x')
       pd_tree = @time PlotData1D(ode_tree.u0, ode_tree.p; curve)
       pd_t8code = @time PlotData1D(ode_t8code.u0, ode_t8code.p; curve)
       pd_p4est = @time PlotData1D(ode_p4est.u0, ode_p4est.p; curve)
       pd_structured = @time PlotData1D(ode_structured.u0, ode_structured.p; curve)

       fig = plot(pd_tree, label = "tree")
       plot!(fig, pd_t8code, label = "t8code")
       plot!(fig, pd_p4est, label = "p4est")
       plot!(fig, pd_structured, label = "structured")
       plot!(fig, legend = true)
       end
  0.009187 seconds (29 allocations: 20.818 MiB)
  0.001079 seconds (19 allocations: 90.828 KiB)
  0.000644 seconds (19 allocations: 90.828 KiB)
  0.254976 seconds (13 allocations: 2.047 MiB)
```