@ranocha 2025-03-31
That was for an initial draft. I you agree with the general approach, I will refactor the code to share more functionality.