@JoshuaLampert 2025-03-31
> That was for an initial draft. I you agree with the general approach, I will refactor the code to share more functionality.

The general approach looks good to me. I think it will be a bit easier to review once both versions share more functionality because then I don't need to compare as much to the t8code version.