@ranocha 2025-03-31
I refactored the code as discussed.